Improved bar display and bug fixes
The exact math for setting the bar display height (volume level) has been
implemented, but for some unknown reason the level is consistently too low.
While investigating this several bugs have been fixed related to TVA
calculations.

A bug in calculating the coarse base pitch, resulting in some instances to
selecting the wrong sample in the break table, is also fixed.

+++ emusc/src/bar_display.h
@@ -71,7 +71,7 @@ private:
   int _peakHold;                            // Off / Type 1 - 3 [1]
 
   QVector<bool> _barDisplay;
-  std::array<float, 16> _currentPartsAmp;
+  std::array<int, 16> _currentPartsAmp;
 
   bool _playAnimation;
   QVector<enum Animation> _animationList;
